requirements.txt
click
pantsbuild.pants==2.12.1
underpants
